plc-reader/requirements.txt
pymcprotocol
python-dotenv
redis
